Add src/component directory and Player.js, move some code to Player.js
