+++ JangSungwoo1901/MusicApp/App.js
@@ -14,15 +14,13 @@ import {
   Image,
   Text,
 } from 'react-native';
-import Slider from '@react-native-community/slider';
 import TrackPlayer from 'react-native-track-player';
 import {
   usePlaybackState,
   useTrackPlayerEvents,
-  useTrackPlayerProgress
 } from 'react-native-track-player/lib';
 import playlistData from './src/data/playlist.json';
-import moment from 'moment';
+import Player from './src/components/Player';
 
 export default function App() {
 
@@ -30,7 +28,7 @@ export default function App() {
   const playbackState = usePlaybackState();
 
   //재생/일시정지 상태변수 선언 
-  const [isPlaying, setIsPlaying] = useState(false);
+  // const [isPlaying, setIsPlaying] = useState(false);
 
   useEffect(() => {
     //TrackPlayer 설정
@@ -53,23 +51,6 @@ export default function App() {
     });
   }, []);
 
-  //track 정보 
-  const [trackTitle, setTrackTitle] = useState("");
-  const [trackArtwork, setTrackArtwork] = useState("");
-  const [trackArtist, setTrackArtist] = useState("");
-  
-  //track 변경 시 동작 
-  useTrackPlayerEvents(["playback-track-changed"], async event => {
-    if (event.type === TrackPlayer.TrackPlayerEvents.PLAYBACK_TRACK_CHANGED) {
-      //변경된 track의 정보를 가져온다.
-      const track = await TrackPlayer.getTrack(event.nextTrack);
-      //track의 정보를 변수에 설정한다.
-      setTrackTitle(track.title);
-      setTrackArtist(track.artist);
-      setTrackArtwork(track.artwork);
-    }
-  });
-
   //track 재생/일시정지 버튼 눌렀을때 동작
   async function togglePlayback() {
     //현재 track 가져오기 
@@ -81,65 +62,29 @@ export default function App() {
       await TrackPlayer.reset();
       await TrackPlayer.add(playlistData);
       await TrackPlayer.play();
-      setIsPlaying(true);
+      // setIsPlaying(true);
     } else{
       //TrackPlayer의 상태
       // STATE_PAUSED -> play() 
       // STATE_PLAY -> pause() 
       if (playbackState === TrackPlayer.STATE_PAUSED) {
         await TrackPlayer.play();
-        setIsPlaying(true);
+        // setIsPlaying(true);
       } else {
         await TrackPlayer.pause();
-        setIsPlaying(false);
+        // setIsPlaying(false);
       }
     }
   }
 
-  //재생/일시정지 버튼
-  function TogglePlaybackButton({ onPress }) {
-    var icon = isPlaying
-      ? require('./src/images/baseline_pause_black_48dp.png')
-      : require('./src/images/baseline_play_arrow_black_48dp.png');
-    return (
-      <TouchableOpacity style={styles.controlButtonContainer} onPress={onPress}>
-        <Image source={icon}></Image>
-      </TouchableOpacity>
-    );
-  }
-
   return (
-    <View style={styles.card}>
-      <Image style={styles.cover} source={{uri:trackArtwork}}></Image>
-      <Text style={styles.title}>{trackTitle}</Text>
-      <Text style={styles.artist}>{trackArtist}</Text>
-      <MusicSlider></MusicSlider>
-        <View style={styles.controls}>
-          <PreviousButton onPress={_skipToPrevious} ></PreviousButton>
-          <TogglePlaybackButton onPress={togglePlayback}></TogglePlaybackButton>
-          <NextButton onPress={_skipToNext}></NextButton>
-        </View>
-    </View>
+    <Player
+      onPrevious={_skipToPrevious}
+      onNext={_skipToNext}
+      onTogglePlayback={togglePlayback}
+    ></Player>
   );
 };
-
-//이전음악으로 이동버튼
-function PreviousButton({ onPress }) {
-  return (
-    <TouchableOpacity style={styles.controlButtonContainer} onPress={onPress}>
-      <Image source={require('./src/images/baseline_skip_previous_black_48dp.png')}></Image>
-    </TouchableOpacity>
-  );
-}
-
-//다음음악으로 이동버튼
-function NextButton({ onPress }) {
-  return (
-    <TouchableOpacity style={styles.controlButtonContainer} onPress={onPress}>
-      <Image source={require('./src/images/baseline_skip_next_black_48dp.png')}></Image>
-    </TouchableOpacity>
-  );
-}
 
 //이전음악으로 이동
 async function _skipToPrevious() {
@@ -154,78 +99,6 @@ async function _skipToNext() {
     await TrackPlayer.skipToNext();
   } catch (_) {}
 }
-
-//재생위치 변경
-_seekTo = async (value) =>{
-  try{
-     await TrackPlayer.seekTo(value);
-  }catch(_){}
-}
-
-//음악 슬라이더
-function MusicSlider() {
-  const progress = useTrackPlayerProgress() 
-  var positionTime = moment(progress.position * 1000).format('mm:ss')
-  var durationTime = moment(progress.duration * 1000).format('mm:ss')
-  return (
-    <View style={styles.progress}>
-      <Slider
-        style={{ width: "100%" }}
-        maximumValue={progress.duration}
-        minimumValue={0}
-        value={progress.position}
-        step={1}
-        onSlidingComplete={(val) => _seekTo(val)}
-          ></Slider>
-      <View style={styles.trackTime}>
-        <Text>{positionTime}</Text>
-        <Text>{durationTime}</Text>
-      </View>
-    </View>
-  );
-}
 const styles = StyleSheet.create({
-  card: {
-    width: "100%",
-    height: "100%",
-    elevation: 1,
-    borderRadius: 4,
-    shadowRadius: 2,
-    shadowOpacity: 0.1,
-    alignItems: "center",
-    shadowColor: "black",
-    backgroundColor: "white",
-    shadowOffset: { width: 0, height: 1 }
-  },
-  controlButtonContainer: {
-    flex: 1,
-    alignItems: 'center',
-  },
-  controls: {
-    flexDirection: 'row',
-    justifyContent: 'center'
-  },
-  cover: {
-    width: 300,
-    height: 300,
-    marginTop: 20,
-    backgroundColor: "grey",
-  },
-  title: {
-    marginTop: 10
-  },
-  artist: {
-    fontWeight: "bold"
-  },
-  progress: {
-    // height: ,
-    width: "90%",
-    marginTop: 10,
-    flexDirection: "column"
-  },
-  trackTime: {
-    width: '100%',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  }
+
 });
